rename unsafe to jit
