+++ .gitignore
@@ -5,4 +5,4 @@ node_modules
 *.log
 
 /dist
-/unsafe
+/jit
