+++ bin/build.js
@@ -53,4 +53,4 @@ function transform(input, outputs) {
 
 transform('src/index.js', pkg.exports['.']);
 console.log('---');
-transform('src/unsafe.js', pkg.exports['./unsafe']);
+transform('src/jit.js', pkg.exports['./jit']);
