+++ src/jit.d.ts
