+++ src/jit.js
