+++ test/jit.js
@@ -1,6 +1,6 @@
 import { test } from 'uvu';
 import * as assert from 'uvu/assert';
-import { hexoid } from '../src/unsafe.js';
+import { hexoid } from '../src/jit.js';
 
 test('exports', () => {
 	assert.type(hexoid, 'function', 'exports function');
